--- SecProject/python/LoOCS_gui/src/dnsmasq.conf
@@ -1,9 +1,0 @@
-interface=wlp3s0mon
-dhcp-range=192.168.0.2,192.168.0.30,255.255.255.0,12h
-dhcp-option=3,192.168.0.1
-dhcp-option=6,192.168.0.1
-server=8.8.8.8
-log-queries
-log-dhcp
-listen-address=127.0.0.1
-port=10467
