--- SecProject/python/LoOCS_gui/src/hostapd.conf
@@ -1,9 +1,0 @@
-interface=wlp3s0mon
-driver=nl80211
-ssid=hackerman1233
-hw_mode=g
-channel=7
-macaddr_acl=0
-ignore_broadcast_ssid=0
-auth_algs=1
-wpa=0
