--- SecProject/python/LoOCS_gui/src/test.py
@@ -1,3 +1,0 @@
-testitest = "wlp3s0mon  IEEE 802.11  Mode:Monitor"
-
-print(testitest.split())
